Move modeling events to new ModelingEvents class
