+++ extensions/ql-vscode/src/model-editor/method-modeling/method-modeling-panel.ts
@@ -7,6 +7,7 @@ import { ModelingStore } from "../modeling-store";
 import { ModelEditorViewTracker } from "../model-editor-view-tracker";
 import { ModelConfigListener } from "../../config";
 import { DatabaseItem } from "../../databases/local-databases";
+import { ModelingEvents } from "../modeling-events";
 
 export class MethodModelingPanel extends DisposableObject {
   private readonly provider: MethodModelingViewProvider;
@@ -14,6 +15,7 @@ export class MethodModelingPanel extends DisposableObject {
   constructor(
     app: App,
     modelingStore: ModelingStore,
+    modelingEvents: ModelingEvents,
     editorViewTracker: ModelEditorViewTracker,
   ) {
     super();
@@ -26,6 +28,7 @@ export class MethodModelingPanel extends DisposableObject {
     this.provider = new MethodModelingViewProvider(
       app,
       modelingStore,
+      modelingEvents,
       editorViewTracker,
       modelConfig,
     );
