+++ extensions/ql-vscode/test/vscode-tests/no-workspace/model-editor/methods-usage/methods-usage-panel.test.ts
@@ -12,6 +12,8 @@ import { ModelingStore } from "../../../../../src/model-editor/modeling-store";
 import { createMockModelingStore } from "../../../../__mocks__/model-editor/modelingStoreMock";
 import { ModeledMethod } from "../../../../../src/model-editor/modeled-method";
 import { Mode } from "../../../../../src/model-editor/shared/mode";
+import { createMockModelingEvents } from "../../../../__mocks__/model-editor/modelingEventsMock";
+import { ModelingEvents } from "../../../../../src/model-editor/modeling-events";
 
 describe("MethodsUsagePanel", () => {
   const mockCliServer = mockedObject<CodeQLCliServer>({});
@@ -33,8 +35,13 @@ describe("MethodsUsagePanel", () => {
       jest.spyOn(window, "createTreeView").mockReturnValue(mockTreeView);
 
       const modelingStore = createMockModelingStore();
+      const modelingEvents = createMockModelingEvents();
 
-      const panel = new MethodsUsagePanel(modelingStore, mockCliServer);
+      const panel = new MethodsUsagePanel(
+        modelingStore,
+        modelingEvents,
+        mockCliServer,
+      );
       await panel.setState(
         methods,
         dbItem,
@@ -51,6 +58,7 @@ describe("MethodsUsagePanel", () => {
   describe("revealItem", () => {
     let mockTreeView: TreeView<unknown>;
     let modelingStore: ModelingStore;
+    let modelingEvents: ModelingEvents;
 
     const hideModeledMethods: boolean = false;
     const mode = Mode.Application;
@@ -65,6 +73,7 @@ describe("MethodsUsagePanel", () => {
       jest.spyOn(window, "createTreeView").mockReturnValue(mockTreeView);
 
       modelingStore = createMockModelingStore();
+      modelingEvents = createMockModelingEvents();
     });
 
     it("should reveal the correct item in the tree view", async () => {
@@ -73,7 +82,11 @@ describe("MethodsUsagePanel", () => {
       });
       const methods = [method];
 
-      const panel = new MethodsUsagePanel(modelingStore, mockCliServer);
+      const panel = new MethodsUsagePanel(
+        modelingStore,
+        modelingEvents,
+        mockCliServer,
+      );
       await panel.setState(
         methods,
         dbItem,
@@ -96,7 +109,11 @@ describe("MethodsUsagePanel", () => {
     it("should do nothing if usage cannot be found", async () => {
       const method = createMethod({});
       const methods = [method];
-      const panel = new MethodsUsagePanel(modelingStore, mockCliServer);
+      const panel = new MethodsUsagePanel(
+        modelingStore,
+        modelingEvents,
+        mockCliServer,
+      );
       await panel.setState(
         methods,
         dbItem,
